Add document renderer for new style PPL NTS summary
